requirements.txt
django==4.1.7
djangorestframework==3.14.0
django-cors-headers==3.14.0
djangorestframework-simplejwt==5.2.2
whitenoise==6.4.0